Breadcrumb for the night. Blech.
Current source of confusion is in `welcome.js.coffee`, in the `createSelection` function starting at line 41. It attempts to use the `setRangeEndpoint` function (lines 37-39) to tell jQuery++’s `$.Range.current().start` and `$.Range.current().end` setters to define the endpoints of a range, which would then be selected via the presently-commented-out `$.Range` constructor call at line 48.

What happens is that we get an internal jQuery++ error, “Uncaught TypeError: Cannot read property ‘setStart’ of undefined”

I’ll pull down the non-minified distribution of jQuery++ tomorrow and see if that sheds any light on the situation. It would be Very Helpful if the `select` function were publicly callable; as it is, it’s internal to the jQuery++ `$.Range` object, and would be called by the (presently commented-out) explicit constructor call at line 48.

I feel soiled.