Learning spec for direct DOM range and selection manipulation. Works.
Hopefully ending the flail that we’ve been going through with various third-party wrappers. Note that this isn’t guaranteed to be as backward-compatible as some of the previous contenders are. WE NO LONGER CARE. If the user is still stuck with Internet Exploder 8 on his desktop, have him borrow an iPad or use his Android Jelly Bean/Kit Kat phone.