Baby steps #1 in a series: csat/TextOffsetValidator is now CsatTextOffsetChecker.
Cleans an imminent-future naming conflict. This also serves as a reminder of the necessity of Trello Card #26 (https://trello.com/c/y3tMryUC/).

Other problems lurk in the Csat* code, but since it’s explicitly for this feature-exploration demo only, not going to worry about them until/unless it’s unavoidable.

Also added Sugar (http://sugarjs.com/) to the Gemfile, as it brings several important fixes to various jQuery objects that we’ll be making use of imminently.