Baby steps #2 and last in a series: csat/NodeIndexValidator is now CsatNodeIndexChecker.
Cleans an imminent-future naming conflict. As previously noted, other problems lurk in the Csat* code, but since it’s explicitly for this feature-exploration demo only, not going to worry about them until/unless it’s unavoidable.

Also left a note in `welcome.js` reminding that the `createSelection` function needs to be completely rewritten, as it depended on the formerly-used jQuery++ `$.Range` plugin. Yes, that’s a design defect masquerading as a change.